[refactor] Split mask editor into smaller files
