+++ src/extensions/core/maskeditor/tools/index.ts
@@ -1,0 +1,3 @@
+export { PaintBucketTool } from './PaintBucketTool'
+export { ColorSelectTool } from './ColorSelectTool'
+export { BrushTool } from './BrushTool'
